refactor(DatePicker): Use hook for setting value
